removing unused rabbit queues from collection exercise
