--- src/main/java/uk/gov/ons/ctp/response/collection/exercise/lib/casesvc.message/sampleunitnotification/package-info.java
@@ -1,11 +1,0 @@
-//
-// This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference
-// Implementation, v2.2.11
-// See <a href="http://java.sun.com/xml/jaxb">http://java.sun.com/xml/jaxb</a>
-// Any modifications to this file will be lost upon recompilation of the source schema.
-// Generated on: 2020.05.02 at 11:47:18 PM BST
-//
-
-@javax.xml.bind.annotation.XmlSchema(
-    namespace = "http://ons.gov.uk/ctp/response/casesvc/message/sampleunitnotification")
-package uk.gov.ons.ctp.response.collection.exercise.lib.casesvc.message.sampleunitnotification;
