+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/message/dto/EventMessageDTO.java
@@ -1,18 +1,18 @@
 package uk.gov.ons.ctp.response.collection.exercise.message.dto;
 
 import lombok.Data;
-import uk.gov.ons.ctp.response.collection.exercise.message.CollectionExerciseEventPublisher;
+import uk.gov.ons.ctp.response.collection.exercise.CollectionExerciseBeanMapper.MessageType;
 import uk.gov.ons.ctp.response.collection.exercise.representation.EventDTO;
 
 @Data
 public class EventMessageDTO {
   private EventDTO event;
 
-  private CollectionExerciseEventPublisher.MessageType messageType;
+  private MessageType messageType;
 
   public EventMessageDTO() {}
 
-  public EventMessageDTO(CollectionExerciseEventPublisher.MessageType messageType, EventDTO event) {
+  public EventMessageDTO(MessageType messageType, EventDTO event) {
     this.event = event;
     this.messageType = messageType;
   }
