+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/representation/SampleUnitChildrenDTO.java
@@ -5,7 +5,6 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.ons.ctp.response.collection.exercise.lib.casesvc.message.sampleunitnotification.SampleUnit;
 
 @Data
 @AllArgsConstructor
