+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/service/change/ScheduledStartDateHandler.java
@@ -4,9 +4,9 @@ import com.godaddy.logging.Logger;
 import com.godaddy.logging.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import uk.gov.ons.ctp.response.collection.exercise.CollectionExerciseBeanMapper.MessageType;
 import uk.gov.ons.ctp.response.collection.exercise.domain.CollectionExercise;
 import uk.gov.ons.ctp.response.collection.exercise.domain.Event;
-import uk.gov.ons.ctp.response.collection.exercise.message.CollectionExerciseEventPublisher;
 import uk.gov.ons.ctp.response.collection.exercise.service.CollectionExerciseService;
 import uk.gov.ons.ctp.response.collection.exercise.service.EventChangeHandler;
 import uk.gov.ons.ctp.response.collection.exercise.service.EventService;
@@ -22,8 +22,7 @@ public final class ScheduledStartDateHandler implements EventChangeHandler {
   @Autowired private CollectionExerciseService collectionExerciseService;
 
   @Override
-  public void handleEventLifecycle(
-      CollectionExerciseEventPublisher.MessageType change, Event event) {
+  public void handleEventLifecycle(MessageType change, Event event) {
     switch (change) {
       case EventCreated:
       case EventUpdated:
