--- src/main/resources/samplesvc/xsd/outbound/sample-unit.xjb
@@ -1,8 +1,0 @@
-<?xml version="1.0"?>
-<jxb:bindings version="1.0" xmlns:jxb="http://java.sun.com/xml/ns/jaxb" xmlns:xjc="http://java.sun.com/xml/ns/jaxb/xjc" jxb:extensionBindingPrefixes="xjc" xmlns:xs="http://www.w3.org/2001/XMLSchema">
-    <jxb:bindings schemaLocation="sample-unit.xsd" node="/xs:schema">
-        <jxb:globalBindings>
-            <xjc:simple />
-        </jxb:globalBindings>
-    </jxb:bindings>
-</jxb:bindings>
